LIBFCREPO-1351. Fix "li" nesting when property action is invalid
Fixed a display issue in which every time an property entry was
not valid, an additional "<li>" stanza was nested inside the outer
<li> stanza. This resulted in the "Property" form slowly expanding, even
to the point of changing the width of the "Properties" column in the
table.

Fixed by adding the "hx-swap" directive with "outerHTML", so that the
`<li class="property">` stanza would be swapped out completely, instead
of being nested inside. This required changing the
"property_detail.html" and "property_form.html" templates to include
the `<li class="property">` stanza.

https://umd-dit.atlassian.net/browse/LIBFCREPO-1351